% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adding_analysis_key}
\alias{adding_analysis_key}
\title{Adds the analysis key to the results table from a create_analysis_x}
\usage{
adding_analysis_key(results)
}
\arguments{
\item{results}{a results table from the adding group_var_value section from create_analysis_x}
}
\value{
results with key
}
\description{
This is a helper to avoid repeating the select in the 5 create_analysis_x functions.
}
\examples{
\dontrun{
adding_analysis_key(results = results)
}
}
\keyword{internal}
