% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adding_analysis_key_ratio}
\alias{adding_analysis_key_ratio}
\title{Adds the analysis key to the results table from a create_analysis_ratio}
\usage{
adding_analysis_key_ratio(results)
}
\arguments{
\item{results}{a results table from the adding group_var_value section from create_analysis_x}
}
\value{
results with key
}
\description{
This is a helper to avoid repeating the select in create_analysis_ratio functions. It differs
as there are 2 analysis variables, the numerator and denominator for the ratios.
}
\examples{
\dontrun{
adding_analysis_key_ratio(results = results)
}
}
\keyword{internal}
