% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adding_group_var_value}
\alias{adding_group_var_value}
\title{Adds the group_var_value to the results table from a create_analysis_x}
\usage{
adding_group_var_value(results, group_var = group_var, grouping_vector)
}
\arguments{
\item{results}{a results table from the calculate section from create_analysis_x}

\item{group_var}{group_var from the create_analysis_x}

\item{grouping_vector}{vector with the variables used in the check the
grouping variable section of a create_analysis_x function}
}
\value{
a dataframe with the results and the group_var_value column added
}
\description{
This is a helper to avoid repeating the select in the 5 create_analysis_x functions.
}
\examples{
\dontrun{
adding_group_var_value(results = results, group_var = group_var, grouping_vector = grouping_c)
adding_group_var_value(results = results, group_var = group_var, grouping_vector = across_by)
}
}
\keyword{internal}
