% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisTools-package.R
\docType{package}
\name{AnalysisTools-package}
\alias{AnalysisTools}
\alias{AnalysisTools-package}
\title{AnalysisTools: A set of support tools for data analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marco Pittarello \email{marco.pittarello1987@gmail.com} (\href{https://orcid.org/0000-0001-6748-8790}{ORCID})

}
\keyword{internal}
