% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_clean_data}
\alias{analysistools_clean_data}
\title{Clean data}
\usage{
analysistools_clean_data
}
\description{
Clean data and Sample Frame
}
\examples{
analysistools_clean_data
analysistools_sample_frame
}
\keyword{datasets}
