% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_dap_with_ratio}
\alias{analysistools_dap_with_ratio}
\title{Example of Data Analysis Plan with ratio}
\usage{
analysistools_dap_with_ratio
}
\description{
Example of Data Analysis Plan with ratio
}
\keyword{datasets}
