% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_loa_with_ratio}
\alias{analysistools_loa_with_ratio}
\title{Example of list of analysis with ratio}
\usage{
analysistools_loa_with_ratio
}
\description{
Example of list of analysis with ratio
}
\keyword{datasets}
