% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_MSNA_template_choices}
\alias{analysistools_MSNA_template_choices}
\title{MSNA template kobo choices (example)}
\usage{
analysistools_MSNA_template_choices
}
\description{
MSNA template kobo choices (example)
}
\keyword{datasets}
