% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_MSNA_template_data}
\alias{analysistools_MSNA_template_data}
\title{MSNA template dataset (example)}
\usage{
analysistools_MSNA_template_data
}
\description{
Clean data and Sample Frame
}
\details{
These data sets include HH data template
}
\examples{
analysistools_MSNA_template_data
}
\keyword{datasets}
