% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_MSNA_template_loa}
\alias{analysistools_MSNA_template_loa}
\title{Example of list of analysis}
\usage{
analysistools_MSNA_template_loa
}
\description{
Example of list of analysis
}
\keyword{datasets}
