% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_MSNA_template_loa_with_ratio}
\alias{analysistools_MSNA_template_loa_with_ratio}
\title{Example of list of analysis with ratio}
\usage{
analysistools_MSNA_template_loa_with_ratio
}
\description{
Clean data and Sample Frame
}
\details{
These data sets include HH data template
}
\examples{
analysistools_MSNA_template_data
}
\keyword{datasets}
