% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_MSNA_template_no_ratio_results_table}
\alias{analysistools_MSNA_template_no_ratio_results_table}
\title{Analysis results example (no ratio)}
\usage{
analysistools_MSNA_template_no_ratio_results_table
}
\description{
Analysis results example (no ratio)
}
\keyword{datasets}
