% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{analysistools_MSNA_template_survey}
\alias{analysistools_MSNA_template_survey}
\title{MSNA template kobo survey (example)}
\usage{
analysistools_MSNA_template_survey
}
\description{
MSNA template kobo survey (example)
}
\keyword{datasets}
