% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arranging_results_columns}
\alias{arranging_results_columns}
\title{Select columns in a specific order}
\usage{
arranging_results_columns(results)
}
\arguments{
\item{results}{a results table from the adding analysis key section from create_analysis_x}
}
\value{
dataframe with the columns in a spefic order
}
\description{
This is a helper to avoid repeating the select in the 5 create_analysis_x functions.
}
\examples{
\dontrun{
arranging_results_columns(results)
}
}
\keyword{internal}
