% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_Spatial_CorrStruct.R
\name{best_Spatial_CorrStruct}
\alias{best_Spatial_CorrStruct}
\title{Find the best spatial correlation structure in GLMM run with 'lme' function in 'nlme' package.}
\usage{
best_Spatial_CorrStruct(
  data,
  X,
  Y,
  var.res,
  range,
  sill,
  nugget,
  spatial.formula
)
}
\arguments{
\item{data}{a data.frame used as data input in the model}

\item{X}{column within the dataset containing the coordinates X (e.g.: dataset$X)}

\item{Y}{column within the dataset containing the coordinates Y (e.g.: dataset$Y)}

\item{var.res}{column within the dataset containing the response variable or model residuals obtained with residuals(model,type="pearson") see @examples}

\item{range}{value of the 'range' identifiable from the semivariogram obtained with (\link[AnalysisTool]{modelCheck_plots}).}

\item{sill}{value of the 'sill' identifiable from the semivariogram obtained with (\link[AnalysisTool]{modelCheck_plots}).}

\item{nugget}{value of the 'nugget' identifiable from the semivariogram obtained with (\link[AnalysisTool]{modelCheck_plots}).}

\item{spatial.formula}{a one sided formula of the form ~ S1+...+Sp, or ~ S1+...+Sp | g, specifying spatial covariates S1 through Sp and, optionally, a grouping factor g. For more details see the 'form' argument of e.g. (\link[nlme]{corGaus}).}
}
\value{
Linear, exponential, gaussian and spherical models plotted on the semivariogram.\cr
It is also provided in the console the string to copy and paste in 'correlation' argument in the functions: (\link[nlme]{gls}), (\link[nlme]{lme}) of 'nlme' package"
}
\description{
When through the function (\link[AnalysisTool]{modelCheck_plots}) spatial autocorrelation is found in the residuals of a model performed
with (\link[nlme]{gls}) or (\link[nlme]{lme}), it may be necessary to specify the structure of this spatial correlation through the 'correlation' argument.\cr
Among the different correlation structures (see (\link[nlme]{CorStruct})), those supported by the present function are:
\itemize{
\item corGaus	 --> Gaussian spatial correlation
\item corExp	 --> Exponential spatial correlation
\item corSpher --> Spherical spatial correlation
\item corLin   --> Linear spatial correlation.\cr
}

See @examples  section for operative details
}
\examples{
1- define from (\link[AnalysisTool]{modelCheck_plots}) which are the values of range, sill, and nugget. 
For a definition of range, sill, and nugget terminology see \href{https://sakai.unc.edu/access/content/group/2842013b-58f5-4453-aa8d-3e01bacbfc3d/public/Ecol562_Spring2012/images/lectures/lecture31/fig1d.png}{here} \cr

2- run 'best_Spatial_CorrStruct' function:
best_Spatial_CorrStruct(data=Dataset, 
X = Dataset$X,
Y = Dataset$Y,
var.res = residuals(model,type="pearson"),
range = 5,
sill = 5,
nugget = 4,
spatial.formula = "~ X + Y")
}
\references{
\itemize{
\item Gałecki, A., & Burzykowski, T. (2013). Linear mixed-effects model. In Linear mixed-effects models using R (pp. 245-273). Springer, New York, NY.
\item \href{https://sakai.unc.edu/access/content/group/2842013b-58f5-4453-aa8d-3e01bacbfc3d/public/Ecol562_Spring2012/docs/lectures/lecture32.htm}{Jack Weiss tutorial (2012). Fitting theoretical models to the empirical semivariogram}
}
}
