% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{char_to_vector}
\alias{char_to_vector}
\title{Turns a string separated by , into a vector}
\usage{
char_to_vector(string)
}
\arguments{
\item{string}{a one length string}
}
\value{
a string vector
}
\description{
This is a helper to avoid repeating the select in the 5 create_analysis_x functions.
}
\examples{
char_to_vector("groupa, groupb, groupc")
}
\keyword{internal}
