% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{check_dap}
\alias{check_dap}
\title{Helper to check the data analysis plan format}
\usage{
check_dap(dap, .design)
}
\arguments{
\item{dap}{a \emph{data analysis plan}}

\item{.design}{Survey design object created with srvyr::as_survey or as_survey_design}
}
\value{
The dap with default values added if missing or an error if something is not correct.
}
\description{
Helper to check the data analysis plan format
}
\examples{
check_dap(.design = srvyr::as_survey(analysistools_MSNA_template_data),
          dap = analysistools_dap)
}
