% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{check_loa}
\alias{check_loa}
\title{Helper to check the list of analysis format}
\usage{
check_loa(loa, design)
}
\arguments{
\item{loa}{a \emph{list of analysis}}

\item{design}{Survey design object created with srvyr::as_survey or as_survey_design}
}
\value{
The loa with default values added if missing or an error if something is not correct.
}
\description{
Helper to check the list of analysis format
}
\examples{
check_loa(
  design = srvyr::as_survey(analysistools_MSNA_template_data),
  loa = analysistools_MSNA_template_loa
)
}
