% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_nan}
\alias{correct_nan}
\title{Helper to correct values to NaN when totals are 0's}
\usage{
correct_nan(
  results,
  stat_columns = c("stat", "stat_upp", "stat_low"),
  total_column = "n_total"
)
}
\arguments{
\item{results}{Results to corrected}

\item{stat_columns}{Columns to be corrected}

\item{total_column}{Total column to be used}
}
\value{
The results with the stats_columns turn to NaN if total column is 0.
}
\description{
Helper to correct values to NaN when totals are 0's
}
\examples{
test_table <- data.frame(stat = c(NaN, 0, 1),
                         stat_low = c(NaN, 0, 1),
                         stat_upp = c(NaN, 0, 1),
                         n_total = c(0,0,1))

correct_nan(test_table)
}
