% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_nan_total_is_0}
\alias{correct_nan_total_is_0}
\title{Helper to correct values to NaN when totals are 0's}
\usage{
correct_nan_total_is_0(
  results,
  stat_columns = c("stat", "stat_upp", "stat_low", "n_w_total", "n_total", "n_w"),
  total_column = "n_total"
)
}
\arguments{
\item{results}{Results to corrected}

\item{stat_columns}{Columns to be corrected}

\item{total_column}{Total column to be used}
}
\value{
The results table with the stats_columns turn to NaN if total column is 0.
}
\description{
Helper to correct values to NaN when totals are 0's
}
\examples{
test_table <- data.frame(
  stat = c(NaN, 0, 1),
  stat_low = c(NaN, 0, 1),
  stat_upp = c(NaN, 0, 1),
  n = c(0, 0, 1),
  n_total = c(0, 0, 1),
  n_w = c(0, 0, 1),
  n_w_total = c(0, 0, 1)
)

correct_nan_total_is_0(test_table)
}
\keyword{internal}
