% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{create_analysis}
\alias{create_analysis}
\title{Create analysis from a design}
\usage{
create_analysis(.design, dap = NULL, group_var = NULL)
}
\arguments{
\item{.design}{Survey design object created with srvyr::as_survey or as_survey_design}

\item{dap}{Data Analysis plan: Default is NULL. If provided it will be used to create the analysis.}

\item{group_var}{Default is NULL. If provided, it will first create a Data Analysis Plan and then
will run the analysis. It should be a vector.}
}
\value{
A list with 3 items:
\itemize{
\item The results table in a long format with the analysis key
\item The dataset that was used
\item The data analysis plan that was used
}
}
\description{
If not \emph{data analysis plan (dap)} is provided, the analysis will run on the overall dataset, and all
the grouping variables set.
}
\details{
The \emph{dap} should contains the following columns :
\itemize{
\item analysis_type: analysis type to be perform. At the moment mean, median, prop_select_one, and
ratio are available.
\item analysis_var: analysis variable to be used as string.
\item group_var: The grouping variable as string. NA if there is no grouping variable. If a combination
of grouping variable should be used together it should be 1 string character separated with a ",".
i.e. c("admin1", "population") and "admin1, population" are different.
\itemize{
\item c("admin1", "population") : will perform the analysis grouping once by admin1, and once by population
\item "admin1, population" : will perform the analysis grouping once by admin1 and admin 2
}
\item level: confidence level to be used. If the column does not exists, .95 will be used.
It can also include a column \strong{level}, if not provided .95 will be set as default.
}

If ratios have to be performed, the \emph{dap} should include the following columns as well:
\itemize{
\item analysis_var_numerator analysis_var_denominator numerator_NA_to_0 filter_denominator_0
}
}
\examples{
create_analysis(.design = srvyr::as_survey(analysistools_MSNA_template_data),
                dap = analysistools_dap)

create_analysis(.design = srvyr::as_survey(analysistools_MSNA_template_data),
                dap = analysistools_dap_with_ratio)

shorter_df <- analysistools_MSNA_template_data[,c("admin1",
                                                  "expenditure_debt",
                                                  "wash_drinkingwatersource")]

create_analysis(.design = srvyr::as_survey(shorter_df),
                group_var = "admin1")
}
