% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_key_table.R
\name{create_analysis_key_table}
\alias{create_analysis_key_table}
\title{Create a table from the analysis_key and results object}
\usage{
create_analysis_key_table(results_table, analysis_key_column = "analysis_key")
}
\arguments{
\item{results_table}{result object with an analysis key}

\item{analysis_key_column}{character vector with the name of the analysis key column.
Default is "analysis key"}
}
\value{
a dataframe with the analysis key and analysis_var_x, analysis_var_value_x, group_var_x,
group_var_value_x, nb_analysis_var and nb_group_var
}
\description{
Helper function to turn a analysis_key to a table. It will return a dataframe with
analysis_var_x, analysis_var_value_x, group_var_x, group_var_value_x, nb_analysis_var,
nb_group_var:
\itemize{
\item Where YY_var_x are the variable names for the dependent and
independent variables.
\item Where YY_var_value_x are the values for the dependent and independent variables.
\item Where x is the number of the variable by order of grouping.
\item nb_xx is the number of variable used in the grouping.
}
}
\examples{
create_analysis_key_table(analysistools_MSNA_template_with_ratio_results_table$results_table)
}
