% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_prop_select_multiple.R
\name{create_analysis_prop_select_multiple}
\alias{create_analysis_prop_select_multiple}
\title{Calculate proportion from a survey design}
\usage{
create_analysis_prop_select_multiple(
  design,
  group_var = NA,
  analysis_var,
  level = 0.95,
  sm_separator = "."
)
}
\arguments{
\item{design}{design survey}

\item{group_var}{dependent variable(s), variable to group by. If no dependent
variable, it should be NA or empty string. If more than one variable, it
should be one string with each variable separated by comma, e.g. "groupa, groupb"
to group for groupa and groupb.
NA is default for no grouping.}

\item{analysis_var}{the independent variable, variable to summarise. It has to be the parent
select multiple}

\item{level}{the confidence level. 0.95 is default}

\item{sm_separator}{Select multiple separator. Default is "."}
}
\value{
a data frame with the proportion for each group and the analysis index.
}
\description{
Calculate proportion from a survey design
}
\examples{
somedata <- data.frame(
  groups = sample(c("group_a", "group_b"), size = 100, replace = TRUE),
  smvar = rep(NA_character_, 100),
  smvar.option1 = sample(c(TRUE, FALSE), size = 100, replace = TRUE, prob = c(.7, .3)),
  smvar.option2 = sample(c(TRUE, FALSE), size = 100, replace = TRUE, prob = c(.6, .4)),
  smvar.option3 = sample(c(TRUE, FALSE), size = 100, replace = TRUE, prob = c(.1, .9)),
  smvar.option4 = sample(c(TRUE, FALSE), size = 100, replace = TRUE, prob = c(.8, .2)),
  uuid = 1:100 \%>\% as.character()
) \%>\%
  cleaningtools::recreate_parent_column(uuid = "uuid", sm_separator = ".")

somedata <- somedata$data_with_fix_concat

create_analysis_prop_select_multiple(srvyr::as_survey(somedata),
  group_var = NA,
  analysis_var = "smvar",
  level = 0.95
)

create_analysis_prop_select_multiple(srvyr::as_survey(somedata),
  group_var = "groups",
  analysis_var = "smvar",
  level = 0.95
)
}
