% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_group_var}
\alias{create_group_var}
\title{Helper to reformat the group_var for the results table}
\usage{
create_group_var(group_var)
}
\arguments{
\item{group_var}{one string with each variable separated by comma, e.g. "groupa, groupb"
to group for groupa and groupb}
}
\value{
group_var separated by " \%/\% " instead of a ","
}
\description{
Helper to reformat the group_var for the results table
}
\examples{
create_group_var("groupa, groupb")
create_group_var("groupa,groupb")
create_group_var(NA)
}
\keyword{internal}
