% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{create_loa}
\alias{create_loa}
\title{Create a list of analysis from design and a grouping variable}
\usage{
create_loa(design, group_var = NULL, sm_separator = ".")
}
\arguments{
\item{design}{Survey design object created with srvyr::as_survey or as_survey_design}

\item{group_var}{The grouping variable as string. If a combination of grouping variable should be
used together it should be 1 string character separated with a ",". i.e. c("admin1", "population")
and "admin1, population" are different:
\itemize{
\item c("admin1", "population") : will perform the analysis grouping once by admin1, and once by population
\item "admin1, population" : will perform the analysis grouping once by admin1 and admin 2
}}

\item{sm_separator}{Separator for choice multiple questions. The default is "."}
}
\value{
a list of analysis.
}
\description{
Helper for the create_analysis. It will take all the columns that are character, logical, double
or integers.
}
\examples{
shorter_df <- analysistools_MSNA_template_data[, c(
  "admin1",
  "expenditure_debt",
  "wash_drinkingwatersource"
)]

create_loa(
  design = srvyr::as_survey(shorter_df),
  group_var = "admin1"
)
}
