% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_loa_from_results.R
\name{create_loa_from_results}
\alias{create_loa_from_results}
\title{Create a list of analysis from a results table}
\usage{
create_loa_from_results(results_table, analysis_key_column = "analysis_key")
}
\arguments{
\item{results_table}{Results table with an analysis key}

\item{analysis_key_column}{character vector with the name of the analysis key column.
Default is "analysis key"}
}
\value{
a list of analysis
}
\description{
Create a list of analysis from a results table
}
\examples{
create_loa_from_results(analysistools_MSNA_template_with_ratio_results_table$results_table)
}
