% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getquintiles.R
\name{getquintiles}
\alias{getquintiles}
\title{Get quintile vector}
\usage{
getquintiles(var, ...)
}
\arguments{
\item{var}{input vector}

\item{...}{use for weights and other arguments see Hmisc package}
}
\value{
quintile vector
}
\description{
Get quintile vector based on input vector
(for optional weights argument see Hmisc package)
}
\examples{
getquintiles(sample(100), weights = rep(1, 100))

}
