% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrollingmean.R
\name{getrollingmean}
\alias{getrollingmean}
\title{Get rolling mean of last n values in vector}
\usage{
getrollingmean(input_vector, n = 3, na.rm = FALSE)
}
\arguments{
\item{input_vector}{variable}

\item{n}{number of values to take mean of}

\item{na.rm}{option to remove missing values before taking mean}
}
\value{
vector with rolling mean of last n (optional: non-missing) values
}
\description{
Get rolling mean of last n values in vector
}
\examples{
getrollingmean(seq(1, 10, 1), 3)
}
