% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getrollingtotal.R
\name{getrollingtotal}
\alias{getrollingtotal}
\title{Get rolling total of last n values in vector}
\usage{
getrollingtotal(input_vector, n = 3, na.rm = FALSE)
}
\arguments{
\item{input_vector}{input vector}

\item{n}{number of values to total}

\item{na.rm}{option to remove missing values before taking mean}
}
\value{
vector with rolling total of last n (optional: non-missing) values
}
\description{
Get rolling total of last n values in vector
}
\examples{
getrollingtotal(seq(1, 10, 1), 3)
}
