% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_color_hue}
\alias{ggplot_color_hue}
\title{Generate ggplot 2 colours}
\usage{
ggplot_color_hue(n)
}
\arguments{
\item{n}{number of colours}
}
\description{
Generates the ggplot2 colours so they can be used in other plots
}
\examples{
ggplot_color_hue(3)
}
