% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_colors}
\alias{ggplot_colors}
\title{Generate ggplot 2 colours}
\usage{
ggplot_colors(n)
}
\arguments{
\item{n}{number of colours}
}
\description{
Generates the ggplot2 colours so they can be used in other plots. Solution taken from https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette/8197703#8197703 and added to this package for easy personal use.
}
\examples{
ggplot_colors(3)
}
