% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{kmeans_elbow}
\alias{kmeans_elbow}
\title{Clustering - generate kmeans elbow plot}
\usage{
kmeans_elbow(data, kmax = 10)
}
\arguments{
\item{data}{dataframe}

\item{kmax}{number of clusters to check}
}
\description{
Generate the elbow plot for a dataframe to understand the number of clusters requirement. Also called scree plot.
}
\examples{
kmeans_elbow(iris[-5])
}
