% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheck_plots.R
\name{modelCheck_plots}
\alias{modelCheck_plots}
\title{Validation plots for models: lm, glm, glmm, gam,gamm}
\usage{
modelCheck_plots(
  model,
  residuals.model,
  database.model,
  response.var,
  explanatory.var,
  coordinate.name = NULL
)
}
\arguments{
\item{model}{a model object, e.g. lm, glm, lme, glmmTMB, gam, lmer, etc}

\item{residuals.model}{e.g. Any kind of "residuals" function able to extract residuals from a modelresiduals(model,type ="pearson")}

\item{database.model}{a data.frame used as data input in the model}

\item{response.var}{the name of response variable between quotes, e.g. "Y"}

\item{explanatory.var}{a vector with listed the explanatory variables used in the model, e.g.
c("X1","X2","X3")}

\item{coordinate.name}{a vector with listed the names of the column reporting the coordinates.
Default is NULL.}
}
\value{
plots for graphical check of model
}
\description{
xx
}
\details{
xx
}
\examples{
xx
}
\references{
xx
}
