% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_analysis.R
\name{review_analysis}
\alias{review_analysis}
\title{Review columns comparing it to another set of columns and spots differences}
\usage{
review_analysis(
  results_table,
  stat_columns_to_review = "stat.x",
  stat_columns_to_compare_with = "stat.y",
  analysis_key_column = "analysis_key",
  prefix = "review"
)
}
\arguments{
\item{results_table}{Results table with an analysis key}

\item{stat_columns_to_review}{Vectors of columns to review (should be paired with
columns_to_compare_with). Default is "stat.x".}

\item{stat_columns_to_compare_with}{Vectors of columns to compare with (should be paired with
columns_to_review). Default is "stat.y"}

\item{analysis_key_column}{character vector with the name of the analysis key column.
Default is "analysis key"}

\item{prefix}{Prefix to be used for the review and comment column. Default is "review".}
}
\value{
A list with two objects:
\itemize{
\item the result table the review and comment columns
\item the review table
}
}
\description{
Wrapper around addindicators::review_variables specific to review analysis.
}
\examples{
analysis_key_column <-  c("mean @/@ income \%/\% NA @/@ NA \%/\% NA",
                          "prop_select_one @/@ water_source \%/\% tap_water @/@ district \%/\% district_a",
                          "prop_select_one @/@ water_source \%/\% tap_water @/@ district \%/\% district_a -/- population \%/\% displaced",
                          "prop_select_multiple @/@ source_information \%/\% relatives @/@ NA \%/\% NA",
                          "ratio @/@ food_expenses \%/\% NA -/- total_expenses \%/\% NA @/@ NA \%/\% NA",
                          "prop_select_one @/@ water_source \%/\% tap_water @/@ population \%/\% displaced",
                          "ratio @/@ food_expenses \%/\% NA -/- total_expenses \%/\% NA @/@ district \%/\% district_a",
                          "prop_select_one @/@ water_source \%/\% tap_water @/@ population \%/\% returnees")
test_analysis_results <- data.frame(
  test = c(
    "test equality",
    "test difference",
    "test Missing in y",
    "test Missing in x",
    "test equality rounding in x",
    "test equality rounding in y",
    "test difference rounding in x",
    "test difference rounding in y"
  ),
  stat_col.x = c(0, 1, 2, NA, 0.00019, 0.0002, 0.00035, 0.0003),
  upp_col.x = c(0, 1, 2, NA, 0.00019, 0.0002, 0.00035, 0.0003),
  stat_col.y = c(0, 2, NA, 3, 0.0002, 0.00019, 0.0003, 0.00035),
  upp_col.y = c(0, 2, NA, 3, 0.0002, 0.00019, 0.0003, 0.00035),
  analysis_key = analysis_key_column
)

review_analysis(test_analysis_results,
                stat_columns_to_review = "stat_col.x",
                stat_columns_to_compare_with = "stat_col.y")
}
