% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Rounds vector}
\usage{
round2(input_vector, digits = 0)
}
\arguments{
\item{input_vector}{numeric vector}

\item{digits}{integer}
}
\value{
rounded numeric vector
}
\description{
Rounds to specified number of places. Rounds 0.5 to
nearest digit, not to even digit as round() does.
Negative digits are allowed.
}
\examples{
round(.5 + -2:4) # IEEE / IEC rounding
round2(.5 + -2:4)
round2(1234, -2)

}
