% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{silhouette_plot}
\alias{silhouette_plot}
\title{Clustering - generate silhouette plot}
\usage{
silhouette_plot(data, kmax = 10)
}
\arguments{
\item{data}{dataframe}

\item{kmax}{number of clusters to check}
}
\description{
Generate a plot to understand the number of clusters using the average silhouette width
}
\examples{
silhouette_plot(iris[-5])
}
