% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_key_table.R
\name{unite_variables}
\alias{unite_variables}
\title{Unite variable from the key_table}
\usage{
unite_variables(key_table)
}
\arguments{
\item{key_table}{a key table built with create_analysis_key_table}
}
\value{
a table with analysis_var, analysis_var_value, group_var, and group_var_value united and
with a \%/\% as separator
}
\description{
Function will unite all analysis_var, analysis_var_value, group_var, and group_var_value from
an key table resulting from create_analysis_key_table.
}
\examples{
\dontrun{
unite_variables(key_table)
}
}
