% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_if_AinB}
\alias{verify_if_AinB}
\title{Checks if all values of a vector is in another vector}
\usage{
verify_if_AinB(.A, .B, msg_error)
}
\arguments{
\item{.A}{Vector to check}

\item{.B}{Reference vector}

\item{msg_error}{Message error to be shown if value are missing}
}
\value{
an error if one is missing
}
\description{
Checks if all values of a vector is in another vector
}
\examples{
\dontrun{
verify_if_AinB(
  c("admin1", "admin2"),
  c("admin1", "pop_group", "uuid", "admin2"),
  "Cannot identify: "
)
}
}
\keyword{internal}
