% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtd.median.R
\name{wtd.median}
\alias{wtd.median}
\title{Get median (for optional weights argument see Hmisc package)}
\usage{
wtd.median(var, ...)
}
\arguments{
\item{var}{input vector}

\item{...}{use for weights and any other arguments see Hmisc package}
}
\value{
median value
}
\description{
Get median (for optional weights argument see Hmisc package)
}
\examples{
wtd.median(seq(1, 100, 1), weights = rep(1, 100))

}
