% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipsave.R
\name{zipsave}
\alias{zipsave}
\title{Zips and saves all files of specified types within specified subdirectories to a different folder}
\usage{
zipsave(
  dirs,
  origin,
  destination,
  pattern = ".R|.Rmd|.Rproj|.html|.css|.yml",
  exceptions = "website/_site",
  keep = 100
)
}
\arguments{
\item{dirs}{vector that lists subfolders to zip}

\item{origin}{folder where the subfolders sit in}

\item{destination}{where to save the zip file}

\item{pattern}{what file types to save (usually small files)}

\item{exceptions}{any subsubfolders to exclude}

\item{keep}{how many zipfiles to keep until the oldest gets deleted}
}
\value{
returns message stating how many files have been zipped and where they have been saved
}
\description{
Zips and saves all files of specified types within specified subdirectories to a different folder
}
\examples{
dir.create("some_test_folders_for_zipsaving")
file.create("some_test_folders_for_zipsaving/file1.R")
zipsave(dirs = "some_test_folders_for_zipsaving",
        origin = getwd(),
        destination = paste0(getwd(), "/zipfiles"))
}
