% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysistools-package.R
\docType{package}
\name{analysistools-package}
\alias{analysistools}
\alias{analysistools-package}
\title{analysistools: Tools to perform some analysis on survey data collected with ODK}
\description{
The package provides a couple of wrappers around srvyr and tidyverse to perform some analysis primarly based on dataset collected throught ODK format.
}
\author{
\strong{Maintainer}: Yann Say \email{yann.say@impact-initiatives.org} (\href{https://orcid.org/0000-0002-7390-4209}{ORCID})

}
\keyword{internal}
