% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis_prop_select_one.R
\name{create_analysis_prop_select_one}
\alias{create_analysis_prop_select_one}
\title{Calculate proportion from a survey}
\usage{
create_analysis_prop_select_one(
  design,
  group_var = NA,
  analysis_var,
  level = 0.95
)
}
\arguments{
\item{design}{design survey}

\item{group_var}{dependent variable(s), variable to group by. If no dependent
variable, it should be NA or empty string. If more than one variable, it
should be one string with each variable separated by comma, e.g. "groupa, groupb"
to group for groupa and groupb.
NA is default for no grouping.}

\item{analysis_var}{the independent variable, variable to summarise}

\item{level}{the confidence level. 0.95 is default}
}
\value{
a data frame with the proportion for each group and th analysis index.
}
\description{
Calculate proportion from a survey
}
\examples{
somedata <- data.frame(
  groups = sample(c("group_a", "group_b"),
    size = 100,
    replace = TRUE
  ),
  value = sample(c("a", "b", "c"),
    size = 100, replace = TRUE,
    prob = c(.6, .4, .1)
  )
)
create_analysis_prop_select_one(srvyr::as_survey(somedata, strata = groups),
  group_var = NA,
  analysis_var = "value",
  level = .95
)
create_analysis_prop_select_one(srvyr::as_survey(somedata, strata = groups),
  group_var = "groups",
  analysis_var = "value",
  level = .95
)
}
